Create README.md
House price prediction using linear regression is a machine learning technique that can be used to estimate the value of a house based on a set of features, such as its size, location, and condition. Linear regression is a supervised learning algorithm, which means that it is trained on a dataset of historical house prices and corresponding features. Once the model is trained, it can be used to predict the price of a new house by providing it with the values of its features.